fix: Resolve all Kotlin compilation errors for v12.0.0-ULTIMATE
- AntiTheftFragment: Use string literal for SIM state changed action
- BatteryOptimizerFragment: Use string literal for LOW_POWER_MODE
- CallRecorderFragment: Fix lateinit on primitive, add Gravity import
- DeviceAdminManagerFragment: Replace deprecated/nonexistent API calls
- DnsChangerFragment: Fix Long to Int type mismatch
- FakeLocationFragment: Use string literal for mock_location
- HoneypotDetectorFragment: Add missing IntentFilter import
- HttpHeaderInspectorFragment: Cast to HttpsURLConnection for sslSession
- IntruderSelfieFragment: Fix Camera.open() with proper front camera ID
- MetadataExtractorFragment: Fix ExifInterface API references
- NotificationHistoryFragment: Add missing Gravity import
- PanicButtonFragment: Add @Suppress for deprecated Browser API
- RamCleanerFragment: Fix unresolved layout reference
- ScreenDimmerFragment: Add missing Intent import
- SensorBoxFragment: Fix type mismatch and null safety
- SslCertificateCheckerFragment: Fix sslSession access and bitLength
- StorageAnalyzerFragment: Make functions suspend for coroutine context
- WhoisLookupFragment: Replace JNDI with DNS-over-HTTPS (Google Public DNS)

+++ app/src/main/java/com/hackerlauncher/modules/DnsChangerFragment.kt
@@ -278,7 +278,7 @@ class DnsChangerFragment : Fragment() {
             }
 
             for ((name, avg) in results) {
-                val bar = "█".repeat((avg / 10).coerceAtMost(30))
+                val bar = "█".repeat((avg / 10).coerceAtMost(30).toInt())
                 val grade = when {
                     avg < 30 -> "FAST"
                     avg < 100 -> "GOOD"
